requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-SocketIO==5.3.4
Flask-Bcrypt==1.0.1
python-socketio==5.10.0
cloudscraper==1.2.71
requests==2.31.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
email-validator==2.1.0
werkzeug==2.3.8
Flask-Migrate==4.0.5
redis==5.0.1
celery==5.3.4
eventlet==0.33.3